requirements.txt
transformers
datasets
peft
bitsandbytes
sentence-transformers
faiss-cpu
torch
accelerate

